backend/requirements.txt
asgiref==3.2.10
certifi==2020.11.8
cloudinary==1.23.0
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-modeladmin-reorder==0.3.1
django-summernote==0.8.11.6
djangorestframework==3.12.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
